39eme commit

+++ src/app/ticket/afficherticketadmin/afficherticketadmin.component.ts
@@ -3,6 +3,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { first } from 'rxjs/operators';
 import { TokenStorageService } from 'src/app/_service/token-storage.service';
 import { TicketService } from 'src/app/_service/ticket.service';
+import { DialogService } from 'src/app/_service/dialog.service';
 
 
 @Component({
@@ -29,7 +30,7 @@ export class AfficherticketadminComponent implements OnInit {
     private TicketService: TicketService,
     private tokenStorageService: TokenStorageService,
     private route: ActivatedRoute,
-    private router: Router) { }
+    private router: Router, private dialogservice: DialogService) { }
 
 
 
@@ -52,12 +53,17 @@ export class AfficherticketadminComponent implements OnInit {
   }
 
   deleteTicket(ticket_id: string) {
-    console.log("sahar" + ticket_id)
-    const ticket = this.tickets.find(x => x.id === ticket_id);
-    this.TicketService.delete(ticket_id)
-      .pipe(first())
-      .subscribe(() => this.tickets = this.tickets.filter(x => x.id !== ticket_id));
-    this.ngOnInit();
+    this.dialogservice.openConfirmDialog('voulez-vous vraiment supprimer ce ticket ?')
+      .afterClosed().subscribe(res => {
+        console.log(res);
+        if (res) {
+          const ticket = this.tickets.find(x => x.id === ticket_id);
+          this.TicketService.delete(ticket_id)
+            .pipe(first())
+            .subscribe(() => this.tickets = this.tickets.filter(x => x.id !== ticket_id));
+          this.ngOnInit();
+        }
+      });
   }
 
   updateStatus2(ticket_id: string) {
